refactor(upstream): relocates code and test files
