+++ exercises/algorithms/src/main/java/com/cbfacademy/sort/App.java
@@ -1,0 +1,7 @@
+package com.cbfacademy.sort;
+
+public class App {
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
